less typing, more any
