+++ defiant-kube/values.yaml
@@ -5,7 +5,7 @@ image:
   name: defiant
   pullPolicy: Always
   # Overrides the image tag whose default is the chart appVersion.
-  tag: "b5526af"
+  tag: "ea78036"
 
 imagePullSecrets: []
 nameOverride: ""
